requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
boto3==1.34.144
botocore==1.34.144
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
dj-database-url==2.2.0
-e git+https://github.com/django/django.git@8c257cecffc314179a5375ab8a16ca0ff3b6fb16#egg=Django
django-anymail==10.3
django-crispy-forms==2.1
django-drf-filepond==0.5.0
django-heroku==0.3.1
django-otp==1.5.0
django-storages==1.14.3
djangorestframework==3.15.1
gunicorn==22.0.0
html5lib==1.1
idna==3.7
jmespath==1.0.1
lxml==5.2.2
markdown2==2.4.13
mysql-connector-python==8.4.0
mysqlclient==2.2.4
oscrypto==1.3.0
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.7.0
xhtml2pdf==0.2.16